Updated testing system

+++ test/run.sh
@@ -1,7 +1,12 @@
 START=$(date +%s)
 
 echo 'Running tests...'
-rm -rf output/*
+
+if ! [ -d output ]; then
+    mkdir output
+else
+    rm -rf output/*
+fi
 
 k=0
 
@@ -16,9 +21,6 @@ do
     projectName=${i:6:(len-12)/2}
     cd $projectName
 
-#    if ! diff ../../$j/package.json package.json; then 
-#        echo '==> FAIL ->' $j/package.json '!==' output/$projectName/package.json
-#        exit 1
     if ! diff ../../$j/make.js .bem/make.js; then
         echo '==> FAIL ->' $j/make.js '!==' output/$projectName/.bem/make.js
         exit 1
